requirements.txt
datasets==2.19.1
transformers==4.41.2
torchmetrics==1.4.0
sacrebleu==2.4.2
matplotlib==3.8.4
numpy==1.26.4
tqdm==4.66.4
